rpm: Switch to CentOS Stream 8
Currently we're taking upcoming RHEL AV contents and semi-manually
rebuilding them inside special COPRs so that they can be used on
Fedora: this process requires additional work, introduces delays
and results in a mix of components that are otherwise not tested
together.

Using CentOS Stream 8 instead of Fedora eliminates all of these
issues: shortly after a new version of QEMU or libvirt is tagged,
the corresponding build will automatically show up in CentOS
Stream 8 and we will be able to consume it; the selection of
packages will also be a lot closer to what Red Hat tests
internally, which should result in an overall more solid base to
build KubeVirt upon.

Using bazeldnf's --nobest option is necessary because CentOS
Stream 8 repositories might contain more than one version of each
package, and we are not necessarily pinning the very latest one.

Note that the image used for some of the tests is still based on
Fedora, at least for now: scsi-target-utils is not available in
CentOS Stream 8, so we'd have to rewrite the corresponding tests
to use an alternative tool such as targetcli. Since this image
is used only for tests, I've decided to leave that challenge for
another day.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>