dockerized: Use CentOS Stream 8 image
This image has been built using the sources as they are in the
previous commit.

When building with the existing image, based on Fedora 32, the
result is a number of errors like

  bazel-out/k8-fastbuild/bin/rpm/libvirt-libs_x86_64/libc-2.28.so:
    error: undefined reference to '_dl_fatal_printf', version 'GLIBC_PRIVATE'

This is surprising because it appears that the symbol in question
is present in CentOS Stream 8's glibc but not Fedora 32's:

  centos-stream-8$ objdump -T /usr/lib64/libc.so.6 | grep -i _dl_fatal_printf
  0000000000000000      DF *UND*	0000000000000000  GLIBC_PRIVATE _dl_fatal_printf
  $

  fedora-32$ objdump -T /usr/lib64/libc.so.6 | grep _dl_fatal_printf
  $

Based on that, I would have expect the opposite scenario to be
problematic - yet here we are.

Tracking down the exact cause for this error and figuring out how
to address it in a satisfactory manner is an ongoing effort.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>